# openapi.yaml
openapi: "3.1.0"
info:
  title: LeanStock API
  version: "2.0.0"
  description: |
    Multi-tenant inventory management system with atomic transfers,
    dead stock decay, predictive reordering, reservation system,
    and full supplier/purchase-order workflows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://leanstock.onrender.com
    description: Production (Render.com)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /reservations:
    post:
      tags: [Reservations]
      summary: Reserve stock with Redis lock
      requestBody:
        required: true
        content:
          application/json:
            example:
              product_id: 1
              location_id: 1
              quantity: 10
              ttl_minutes: 30
      responses:
        "201":
          description: Reservation created
        "409":
          description: Insufficient available stock or lock conflict
    get:
      tags: [Reservations]
      summary: List reservations
      parameters:
        - name: product_id
          in: query
          schema:
            type: integer
        - name: location_id
          in: query
          schema:
            type: integer
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Reservation list
  /reservations/{reservation_id}:
    delete:
      tags: [Reservations]
      summary: Release reservation
      parameters:
        - name: reservation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Released
...
  /suppliers:
    get:
      tags: [Suppliers]
      summary: List active suppliers
      responses:
        "200":
          description: Supplier list
    post:
      tags: [Suppliers]
      summary: Create supplier (manager+)
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "ТОО Снабжение"
              email: "orders@supplier.com"
              phone: "+7 777 123 4567"
              lead_time_days: 7
      responses:
        "201":
          description: Created
  /suppliers/{supplier_id}:
    get:
      tags: [Suppliers]
      summary: Get supplier
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier details
    put:
      tags: [Suppliers]
      summary: Update supplier (manager+)
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              lead_time_days: 5
      responses:
        "200":
          description: Updated
    delete:
      tags: [Suppliers]
      summary: Soft-delete supplier (manager+)
      parameters:
        - name: supplier_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deactivated

  # ── Purchase Orders ────────────────────────────────────────────────────────
...